+++ extensions/uuid-generator/src/generateV5.tsx
@@ -9,6 +9,7 @@ interface UUIDV5Arguments {
 interface Preferences {
   upperCaseLetters: boolean;
   uuidNamespace: string;
+  defaultAction: string;
 }
 
 // don't want to cause a heap error, so cap it 😱
@@ -17,7 +18,7 @@ const UUID_MAX_NUMBER = 10000;
 export default async (props: { arguments: UUIDV5Arguments }) => {
   let { numberOfUUIDsToGenerate } = props.arguments;
   const { name } = props.arguments;
-  const { upperCaseLetters, uuidNamespace } = getPreferenceValues<Preferences>();
+  const { upperCaseLetters, uuidNamespace, defaultAction } = getPreferenceValues<Preferences>();
 
   if (!numberOfUUIDsToGenerate) {
     numberOfUUIDsToGenerate = "1";
@@ -40,9 +41,16 @@ export default async (props: { arguments: UUIDV5Arguments }) => {
       if (upperCaseLetters) {
         uuids = uuids.map((element) => element.toUpperCase());
       }
-      const successMessage = uuids.length > 1 ? `Copied ${uuids.length} new UUIDs.` : `Copied new UUID: ${uuids}`;
-      await Clipboard.copy(uuids.join("\r\n"));
+    
+      if (defaultAction === "copy") {
+        await Clipboard.copy(uuids.join("\r\n"));
+      } else if (defaultAction === "paste") {
+        await Clipboard.paste(uuids.join("\r\n"));
+      }
+      const action = defaultAction === "copy" ? "Copied" : "Pasted";
+      const successMessage = uuids.length > 1 ? `${action} ${uuids.length} new UUIDs.` : `${action} new UUID: ${uuids}`;
       await showHUD(`✅ ${successMessage}`);
+
     } else {
       await showToast({
         style: Toast.Style.Failure,